tool re-org
