+++ rust/crates/sift_mcp/src/tool/mod.rs
@@ -1,5 +1,9 @@
+pub mod assets;
+pub mod channels;
+pub mod common;
 pub mod data;
 pub mod explore;
-pub mod list;
 pub mod ping;
+pub mod reports;
 pub mod rules;
+pub mod runs;
